requirements.txt
fuzzyset==0.0.19
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
Pillow==7.2.0
PyYAML==5.3.1
scipy==1.5.2
tqdm==4.50.2
pyueye==4.95.0
minio==7.1.11
python-crontab==2.6.0
defusedxml==0.7.1
paramiko==2.11.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
torch==1.10.1
torchvision==0.11.2
cx_oracle==8.3.0
scikit-learn==0.24.2
seaborn==0.11.2
google-cloud-vision==2.7.3
pysftp==0.2.9
IPython==7.16.3
impyla==0.18.0